[stat_visualization mas_user_info] 出力する処理の順番を変更 (#710)
* visualize statistics: コマンドラインの説明を修正

* herupuメッセージの"--help"を"global optional arguments"に移動した

* masterブランチをmainプランチに名前を変えたときの修正

* ログローテとの修正

* ヘルプメッセージの修正

* version uup

* update

* poetyru pdate

* ヘルプメッセージの修正

* 入力チェック

* poetry update

* format

* format

+++ annofabcli/stat_visualization/mask_visualization_dir.py
@@ -69,10 +69,6 @@ def mask_visualization_dir(
     )
     print_csv(masked_df_member_performance, output=str(output_dir / FILENAME_PERFORMANCE_PER_USER))
 
-    df_task = pandas.read_csv(str(project_dir / FILENAME_TASK_LIST))
-    _replace_df_task(df_task, replacement_dict_by_user_id=replacement_dict_by_user_id)
-    print_csv(df_task, output=str(output_dir / FILENAME_TASK_LIST))
-
     # メンバのパフォーマンスを散布図で出力する
     write_performance_scatter_per_user(output_dir / FILENAME_PERFORMANCE_PER_USER, output_dir=output_dir / "scatter")
 
@@ -80,23 +76,40 @@ def mask_visualization_dir(
     if exclude_masked_user_for_linegraph:
         user_id_list = list(not_masked_user_id_set)
 
-    # メンバごとにパフォーマンスを折れ線グラフで出力する
-    write_linegraph_per_user(
-        output_dir / FILENAME_TASK_LIST,
-        output_dir=output_dir / "line-graph",
-        minimal_output=minimal_output,
-        user_id_list=user_id_list,
-    )
+    task_csv_file = project_dir / FILENAME_TASK_LIST
+    if task_csv_file.exists():
+        df_task = pandas.read_csv(str(project_dir / FILENAME_TASK_LIST))
+        _replace_df_task(df_task, replacement_dict_by_user_id=replacement_dict_by_user_id)
+        print_csv(df_task, output=str(output_dir / FILENAME_TASK_LIST))
+    else:
+        logger.warning(f"'{task_csv_file}' が存在しないため、" f"'{output_dir / FILENAME_TASK_LIST}' は出力しません。 ")
 
-    df_worktime = pandas.read_csv(str(project_dir / "ユーザ_日付list-作業時間.csv"))
-    df_masked_worktime = create_masked_user_info_df(
-        df_worktime,
-        not_masked_biography_set=not_masked_biography_set,
-        not_masked_user_id_set=not_masked_user_id_set,
-    )
-    worktime_per_date_obj = WorktimePerDate(df_masked_worktime)
-    worktime_per_date_obj.plot_cumulatively(output_dir / "line-graph/累積折れ線-横軸_日-縦軸_作業時間.html", user_id_list)
-    worktime_per_date_obj.to_csv(output_dir / "ユーザ_日付list-作業時間.csv")
+    if (output_dir / FILENAME_TASK_LIST).exists():
+        # メンバごとにパフォーマンスを折れ線グラフで出力する
+        write_linegraph_per_user(
+            output_dir / FILENAME_TASK_LIST,
+            output_dir=output_dir / "line-graph",
+            minimal_output=minimal_output,
+            user_id_list=user_id_list,
+        )
+
+    user_date_csv_file = project_dir / "ユーザ_日付list-作業時間.csv"
+    if user_date_csv_file.exists():
+        df_worktime = pandas.read_csv(str(user_date_csv_file))
+        df_masked_worktime = create_masked_user_info_df(
+            df_worktime,
+            not_masked_biography_set=not_masked_biography_set,
+            not_masked_user_id_set=not_masked_user_id_set,
+        )
+        worktime_per_date_obj = WorktimePerDate(df_masked_worktime)
+        worktime_per_date_obj.plot_cumulatively(output_dir / "line-graph/累積折れ線-横軸_日-縦軸_作業時間.html", user_id_list)
+        worktime_per_date_obj.to_csv(output_dir / "ユーザ_日付list-作業時間.csv")
+    else:
+        logger.warning(
+            f"{user_date_csv_file}が存在しないため、"
+            f"'{output_dir / 'line-graph/累積折れ線-横軸_日-縦軸_作業時間.html'}', "
+            f"'{output_dir / 'ユーザ_日付list-作業時間.csv'}' は出力しません。"
+        )
 
 
 def main(args):
